requirements.txt
selenium
paho-mqtt==2.1.0
beautifulsoup4
